refactor(Notification): "DDD 구조로 변환"
