--- src/main/java/com/github/nenidan/ne_ne_challenge/domain/notification/dto/request/ReadNotificationRequest.java
@@ -1,12 +1,0 @@
-package com.github.nenidan.ne_ne_challenge.domain.notification.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class ReadNotificationRequest {
-	boolean isRead;
-}
