--- src/main/java/com/github/nenidan/ne_ne_challenge/domain/notification/enttiy/Notification.java
@@ -1,52 +1,0 @@
-package com.github.nenidan.ne_ne_challenge.domain.notification.enttiy;
-
-import com.github.nenidan.ne_ne_challenge.domain.user.entity.User;
-import com.github.nenidan.ne_ne_challenge.global.entity.BaseEntity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class Notification extends BaseEntity {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	private String title;
-	private String content;
-
-	@Enumerated(EnumType.STRING)
-	private NotificationType type;
-
-	@Setter
-	private boolean isRead;
-
-	@ManyToOne
-	@JoinColumn(name = "receiver_id")
-	private User receiver;
-
-	@ManyToOne
-	@JoinColumn(name = "sender_id")
-	private User sender;
-
-	public Notification(String title, String content, NotificationType type, User receiver, User sender) {
-		this.title = title;
-		this.content = content;
-		this.type = type;
-		this.receiver = receiver;
-		this.sender = sender;
-	}
-}
