--- src/main/java/com/github/nenidan/ne_ne_challenge/domain/notification/exception/NotificationException.java
@@ -1,10 +1,0 @@
-package com.github.nenidan.ne_ne_challenge.domain.notification.exception;
-
-import com.github.nenidan.ne_ne_challenge.global.exception.BusinessException;
-import com.github.nenidan.ne_ne_challenge.global.exception.ErrorCode;
-
-public class NotificationException extends BusinessException {
-  public NotificationException(ErrorCode errorCode) {
-    super(errorCode);
-  }
-}
