+++ src/main/java/com/github/nenidan/ne_ne_challenge/notification/domain/vo/ReceiverInfo.java
@@ -1,0 +1,14 @@
+package com.github.nenidan.ne_ne_challenge.notification.domain.vo;
+
+import jakarta.persistence.Embeddable;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Embeddable
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class ReceiverInfo {
+	private Long id;
+}
