+++ src/main/java/com/github/nenidan/ne_ne_challenge/notification/infra/NotificationSender.java
@@ -1,0 +1,5 @@
+package com.github.nenidan.ne_ne_challenge.notification.infra;
+
+public interface NotificationSender {
+	void send(Long userId, String title, String content);
+}
